bring back layers panel
